Extract a common ThreadDataManager for use by both bookmarks and threads lists.
